Fix test ocaml-general after collection changes exposed an archive format error.
A change in the collection scripts when dropping IPC::Run exposed an
archive error that was previously ignored. The error is fixed here.
Now the package builds a valid archive named libbaz.a.

Also devolves d/rules, which created files on the go, into static
files installed via debhelper. Most of them are in orig/ocaml.

Also, while 'touch' created files with zero size, the new files have
content.  That makes it possible for Lintian to conclude that
no-architecture dependent files are present. The resulting tag was
added to the expected tags.

Gbp-Dch: ignore