For test preparation, use multiplexing from IO::Async::Function instead of ::Routine/::Channel combo.
Uses the multiple worker feature offered by IO::Async::Function
instead of the homegrown solution via ::Routine and ::Channel. Makes
the code a lot shorter.

Due to unfavorable interaction with Path::Tiny, we only use test
workers once. The reason is unclear, but it looks like the locks are
not being removed because other processes keep them open. The issue
was documented at Path::Tiny, but is probably because IPC::Run is used
in some places in Lintian:

https://github.com/dagolden/Path-Tiny/issues/224

The module Test::State is no longer used and can now be deleted.

Gbp-Dch: ignore