Bundle calls to Path::Tiny->append_utf8 to avoid locking problems while using IO::Async::Function.
The attempt to use IO::Async::Function and its worker process
implementation in runtests instead of our home-grown solution
encountered a number of challenges with Path::Tiny. That module locks
files with flock, and in the multi-process environment, the locks did
not get released properly.

This bundles the calls to append_utf8 to avoid successive calls from
blocking when the lock is not released, but there may be additional
problems. The issue is further documented here:

https://github.com/dagolden/Path-Tiny/issues/224

Additional solutions appear to bracked the calls with setting and
unsetting PERL_PATH_TINY_NO_FLOCK or explicitly inserting 'flock($fh,
LOCK_UN);' before the close here:

https://metacpan.org/source/DAGOLDEN/Path-Tiny-0.108/lib/Path/Tiny.pm#L551

Gbp-Dch: ignore