In collection/java-info, replace routines based on IPC::Run with IO::Async.
Special care was taken to replicate existing code, except that indices
are processed in memory instead of via file handles. The processes are
short-lived, so any extra memory consumed should be freed quickly.

This commit is part of an effort to drop IPC::Run from Lintian.

The framework of choice for Lintian is now IO::Async. It performs
well, but has its own pitfalls: For example, the SIGCHLD handler is
replaced in the parent, which means that other forms of spawning
processes may hang.  Also, until version 0.73, nested usage required
to undefine the loop singleton $IO::Async::Loop::ONE_TRUE_LOOP.

Overall, it is hoped that this change will lead to code simplification
in the collects. It may also contribute to greater stability on
lindsay.

Gbp-Dch: ignore