In t/scripts/spellintian, replace routines based on IPC::Run.
The previous trick to silence debug messages related to IPCRUNDEBUG is
no longer needed. The test now prefers the libraries from TEST_ROOT,
which no longer load IPC::Run. The conflict with IPC::Run3 is gone.
There are no more spurious debug messages.

This commit is part of an effort to migrate Lintian to IO::Async for
multi process control. A previous commit series did something similar
for collections. The present series replaces IPC::Run in the remainder
of Lintian proper.

The sole exception is the library Lintian::Command, which is now used
only by the reporting code. That module will be deleted after the new
reporting code, which also uses IO::Async, is tested and merged.

This series does not address the reporting code.

Gbp-Dch: ignore