For stretch-backports, also Build-Depend on XS version of Text::CSV in addition to the pure Perl version.
For some time, Jenkins has produced CI failures for stretch-backports.
This should fix it.

Text::CSV comes with two backends. One is XS; the other pure Perl. The
XS version is enabled by default but packaged separately. As a
recommended package, the XS version is not automatically installed by
the default resolver on Jenkins (presumably sbuilds's default apt),
but is installed when manually using sbuild with the option
'--build-dep-resolver=aptitude' (which was required for the backport
repo to work). The issue was hard to reproduce.

The error occurs because, for the version in stretch-backports
(1.33-2), the two backends behave differently.

The pure Perl implementation does not allow an 'escape_char' before a
field separator. It triggers the error messsage "EUF" seen in the logs
(internal code 4004).  (As a side note, the 'escape_char' seemed
primarily intended to escape the 'quote_char', and not the field
separator.)

Modern versions of the module no longer use the EUF error message or
the 4004 code.  More significantly, the Perl version currently in
testing (1.99-1) is not so affected.