For stretch-backports, fix test for 'rules-requires-root-implicitly'.
In stretch, control tarballs are compressed with gz, while later
releases use xz. The test for the tag 'rules-requires-root-implicitly'
checks all classification tags, including one for the compression
format. The tag is not needed, but causes failure in
stretch-backports.  Limits the classification tags to the check
'control-file'.

Causes the test to pass on stretch-backports, as well.