Add collection helper functions based on IO::Async.
This will help with the gradual migration of the collection scripts to
IO::Async based process management. Provides the following functions:

    safe_qx:

The existing helper function safe_qx has to be replaced gradually. It
may not be mixed with IO::Async (although this new function can). The
new function, on the other hand, cannot be mixed with older code that
still uses IPC::Run.

The new function is located in lib/Lintian/Util.pm and not, like the
old one, in lib/Lintian/Command.pm. The plan is to get rid of the
latter after a transition period.

    gzip:

Takes data from memory, such as an uncompressed file index, and stores
it in a gzip-compressed file.

    gunzip_file

Unzips a file and stores it in another.

    sort_file_index:

Sorts a file index in memory. Commonly used function in the collection
scripts.

This commit is part of an effort to drop IPC::Run from Lintian.

The framework of choice for Lintian is now IO::Async. It performs
well, but has its own pitfalls: For example, the SIGCHLD handler is
replaced in the parent, which means that other forms of spawning
processes may hang.  Also, until version 0.73, nested usage required
to undefine the loop singleton $IO::Async::Loop::ONE_TRUE_LOOP.

Overall, it is hoped that this change will lead to code simplification
in the collects. It may also contribute to greater stability on
lindsay.

Gbp-Dch: ignore