Fix test deb-format-extra-underscore-member-before-control.
Adjusts this test to use the new and improved 'debs' skeleton. It
features a make-based builder that works with all tests. It no longer
uses t/bin/create-deb.

Control files are located in the ./DEBIAN folder. The mandatory
documentation files copyright and changelog are in the ./doc folder.
Extra data files are provided in the ./root folder.  The package is
built from files in the ./build folder, which is also the folder in
which the scripts tend to operate to set links or permissions, for
example.

Gbp-Dch: ignore