Move templates for mandatory docs to their own template set.
Like the previous commit, this changes accounts properly for the fact,
that the mandatory documentation items in binary files are not part of
the 'control' set (the DEBIAN folder, when working with dpkg-deb).

In the new template, these files are installed in ./doc. The old
control set now goes into ./DEBIAN.

Gbp-Dch: ignore