From the old suite debs, split up two tests for two checks each, and move them.
Two tests that were each associated with exactly two checks were
split. A copy was moved to each check. The following tests were
affected:

    md5sums-traversal
    unpack-numerical-owner

The fields for Test-Against were likewise split such that the tags
mentioned there belong only to the check with which that copy of the
test is associated.

Gbp-Dch: ignore