From the old suite debs, add test deb-format-record-size to check deb-format.
This test looks for false positives (i.e. emits no tags) but did not
specify a Test-Against. Based on the test name and the tar features
used, the test was associated with the check deb-format. A field Check
was added to the test specification.

The test was also moved to the appropriate folder.

Gbp-Dch: ignore