From the old suite debs, add test control-field-traversal-4 to check control-files.
This test looks for false positives (i.e. emits no tags) but did not
specify a Test-Against. Based on the test name and the files in it,
the test was associated with the check 'control-files'. A field Check
was added to the test specification.

The test was also moved to the appropriate folder.

The test may be misnamed (fields vs. files) and may be renamed in the
future.

Gbp-Dch: ignore