From the old suite changes, move all tests related to a single check to that check.
Thirteen tests that were associated with exactly one check were moved
to that check. The old suite 'changes' will soon be fully integrated
into t/tags/checks. The other tests only use a different skeleton.

The old behavior get be obtained by selecting for the skeleton when
running tests. This is a sample command:

	t/bin/runtests --onlyrun=skeleton:changes

The 'skeleton' selector was previously added to the test apparatus.

Gbp-Dch: ignore