From the old suite 'changes', add test changes-directory-traversal to check changes-file.
This test looks for false positives (i.e. emits no tags) but did not
specify a Test-Against. Based on the test name and the files in it,
the test was associated with the check 'changes-file'. A field Check
was added to the test specification.

The test was also moved to the appropriate folder.

Gbp-Dch: ignore