Remove README for the old suite 'changes'.
The organization of the test suite is still in flux. Documentation
will be provided when things are settled down. For now, please examine
the skeleton, the templates and the builder to see how the test is
build. You can also see all the files produced in ./debian/test-out/.

Gbp-Dch: ignore