Allow symlinks in autopkgtest paths.
The check 'testsuite' uses Lintian::Collect::Package::index to
ascertain whether files exist. That method does not allow symbolic
links. Calls index_resolved_path instead.

This issue arose when the test 'debian-symlink' from the old suite
'source' was adapted to a new builder. There, the tag
'inconsistent-testsuite-field' would appear if a Testsuite field were
present (or if standard templates are used). This test will use the
standard templates in the future, and should not trigger the tag.

In a twist, fixing the bug also triggers the tag in that test.  That
is because the check now recognizes the tests, but there is no field
Testsuite in the dsc. The check emits the 'inconsistent' tag via xor
when one is true but not the other.

It would probably make sense to split the tag to alert the user more
meaninfully to either the presence of a surplus specification or the
lack of a declaration in the dsc.

While it is not common to have links in or around ./debian, the test
'debian-symlink' appears to be designed specifically as a false
positive for that scenario.