Adjust test 'debian-symlink' after allowing symlinks in testsuite paths.
Adds a tag triggered by a recent change in check 'testsuite'.

The check 'testsuite' used to call Lintian::Collect::Package::index to
ascertain whether files exist. That method dit not allow symbolic
links and was recently changed to index_resolved_path.

This issue arose originally when this test was adapted to use a new
builder and standard templates. The tag 'inconsistent-testsuite-field'
appeared. The test seemed to be designed to detect specifically such a
situation.

In a twist, fixing the bug also triggers the tag, which is added here.
That is because the check now recognizes the tests, but there is no
field Testsuite in the dsc. (The check emits the 'inconsistent' tag via
xor when one is true but not the other.)

The new tag also looks a little bit like a duplicate next to the
existing tag 'testsuite-autopkgtest-missing'. Perhaps the
'inconsistent' tag should be removed. The lack of a declaration is
already handled by the tag 'testsuite-autopkgtest-missing'.

Gbp-Dch: ignore