Delete test control-file-syntax-error; tag cannot be reproduced using dpkg-source.
The tag 'syntax-error-in-control-file' cannot be tested using standard
tools. The conditions that would be necessary cause dpkg-source to
abort with the message:

dpkg-source: error: syntax error in control-file-syntax-error-1/debian/control at line 11: line with unknown format (not field-colon-value)

The issue arose when the test was converted to standard tools. (It
previously used pattern substitution.) Since most people use standard
tools, it makes little sense to pursue the test further.

Removes the test. If there is no other way to test the tag, the tag
may eventually also be removed from Lintian.

Gbp-Dch: ignore