To fix stable-bpo, require a minimum version of nodejs in test pkg-js-tools-test-is-missing.
For some reason the Gitlab CI pipeline for stable-bpo [1] uses nodejs
4.8.2~dfsg-1, even though the backports contain 8.11.1~dfsg-2~bpo9+1.
The newer version is apparently required to use the buildsystem
'nodejs' in debhelper. Restricts the version in the failed test.
Fixes the build error and restores the Gitlab pipeline for stable-bpo.

Does not address the question why Gitlab prioritizes the older version
of nodejs despite these settings in .gitlab-ci.yaml:

    env DEBIAN_FRONTEND=noninteractive apt-get -q -y \
        --solver aspcud -o APT::Solver::Strict-Pinning=0 \
        -o Debug::pkgProblemResolver=yes build-dep .

The exact message from the failed test
pkg-js-tools-test-is-missing is:

    dh build --with nodejs
       dh_update_autotools_config
       dh_autoreconf
       dh_auto_test --buildsystem=nodejs
          /usr/bin/node -e require\(\".\"\)
    Can't exec "/usr/bin/node": No such file or directory at /usr/share/perl5/Debian/Debhelper/Dh_Lib.pm line 475.
    dh_auto_test: /usr/bin/node -e require\(\".\"\) failed to execute: No child processes
    dh_auto_test: /usr/bin/node -e require\(\".\"\) returned exit code 10
    debian/rules:4: recipe for target 'build' failed

[1] https://salsa.debian.org/lintian/lintian/-/jobs/156352

Gbp-Dch: ignore