Fix occasional hangs during cleanup of processables by switching to IO::Async::Function.
Commit 8f8ddc1d should only wait on futures but not call $loop->run,
as it may cause the cleanup to hang. Rewritten using
IO::Async::Function.

Unfortunately, this solution currenty requires the check for unreaped
processes in commands/lintian.pm to be turned off. The exact reason is
uncertain, but it is very likely related to IO::Async keeping a worker
process alive. This issue should be revisited. Also see 3bbcc3b.

Gbp-Dch: ignore