Add default match strategy 'tags' for tests; set default output format to 'universal'; drop sort.
Adjusts the test case defaults to make the matching strategy for
'tags', which was previously used in all tests, the explicit default.
This is done to so that another matching strategy, 'literal', can be
introduced. The second strategy will compare lintian output literally,
without any processing or sorting. It is useful, for example, when
comparing warning messages not in tag format.

Also sets the default output format to 'universal'. This format is
newly available and saves calls to several Perl scripts in the test
runner. Hopefully, that will shorten the test execution times.

Tags in the 'universal' format are always sorted. It generally does
not make sense to test for the order in which tags appear. The sorting
should really be done in Lintian for all output formats. It is no
longer available in the test runner.

Gbp-Dch: ignore