Implement literal match strategy in test runner; adjust for universal tag format; drop sorting.
This provides for the new mechanisms in the test runner. The tag
processing is now done inside the runner, which saves three calls to
Perl scripts: two calls to tagextract and one call to tagdiff.

This step is thought to save about seven minutes in an average
Gitlab run (but the evidence is still unclear).

Gbp-Dch: ignore