Implement universal output format in Lintian.
This is a new output option that shortcuts to the output of
tagextract. This was done to avoid processing tags during the test, to
save time. This format is a tag-oriented, common subset of all the
other output formats. It is well-suited to evaluate how well a
particular check is working.

Among the notable features is the absence of the severity tag, which
has no bearing on the effectiveness of a check, as well as that the
output is always sorted with the packages on a particular order. For
example, source packages always come before binary packages.

The format looks like this:

scripts (source): binary-arch-rules-but-pkg-is-arch-indep
scripts (source): ancient-standards-version 3.2.1 (released 2000-08-24) (current is CURRENT)
scripts (binary): wrong-path-for-interpreter usr/bin/rubyfoo (#!/bin/ruby1.8 != /usr/bin/ruby1.8)
scripts (binary): wrong-path-for-interpreter usr/bin/lefty-foo (#!/usr/local/bin/lefty != /usr/bin/lefty)
scripts (binary): unusual-interpreter usr/bin/suidperlfoo #!/usr/bin/suidperl