Adjust tag filter script for test runtests-calibration.
The tag filter script for this test previously worked only for the EWI
output format, but the test now uses the universal output format.
Adjusts the script to work with the new, univeral format instead.

The tag filter process was previously known as 'calibration'. We now
use that name when calibrating the expected tags in tests after
changes to the Lintian checks have been made. These scripts will
hopefully be renamed soon. The use of the word calibration for this
particular action should fade into disuse.

Gbp-Dch: ignore