Convert expected tags for test runtest-calibration from EWI to universal format.
Ordinarily, it would not matter which matching strategy this test
employs, but the test filtering capability (previously known as
calibration), which is being tested here, is only available in the
'tags' strategy. That strategy requires the expected tags to be in
universal format. Converts the tags for that test.

This feature will probably be tested in a harness test in the future.

Gbp-Dch: ignore