Turn off check for unreaped processes at the end of Lintian run.
While the cleanup via IO::Async::Function is very elegant, it seems to
keep an unreaped process around. The exact reasons are still under
investigation together with #io-async.

The check should be enabled again after all of Lintian was converted
to IO::Async.  For that reason, it was only commented out with an
explanation.

Gbp-Dch: ignore