Depend and Build-Depend on libio-async-loop-epoll-perl (>= 0.20); limits to buster and later.
This commit limits Litian to buster and later, due to the availability
of this library.

This loop performs better than the standard IO::Async::Loop::Poll. It
also lacks some bugs.

Unfortunately, I cannot remember the details; some files may not have
been closed properly. The symptom is that the tests run fine until the
very end when some processes get stuck. The test suite then hangs.
Child processes using IO::Async will not work properly with the
standard IO::Async::Loop::Poll, which is probably why the Eloop is
recommended in APT.

I believe, but am no longer sure, that this also cures a problem with
flocks not being released in Path::Tiny. More information on that is
in commit 0e84da82.

This commit is required to convert Lintian away from IPC::Run and to
IO::Async.

Signed-off-by: Chris Lamb <lamby@debian.org>