In Lintian::Lab::Entry, use Path::Tiny instead of system() for mkdir.
The modules in IO::Async replace the SIGCHLD handler and do not work
well together with other forms of process management. That also
applies to system() when used in the parent while a loop is active, as
might happen with lab entries. Replaces the call to system() with a
call to Path::Tiny, as a precaution.

Gbp-Dch: ignore