Let tagadjust operate on universal tag files.
No longer extracts universal tags from the tag file given on the
command line. Instead assumes the file is in universal format already.
Likewise, any edits for purposes of calibration are no longer made in
the EWI format but directly in the universal tag format.

Also sorts the output in the preferred order for universal tags, since
they are now all sorted in the test specifications.

Gbp-Dch: ignore