Switch exec branch in Lintian::CollScripts->collect to an invocation that does not use IO::Async.
For some reason IO::Async::Loop->run_child hangs when setting the collection
scripts to Interface: exec. (Maybe ONE_TRUE_LOOP must be undefined for proper
nesting as fixed upstream in 0.73, but that cannot be tested when invoking a
command instead of a code snippet.) The exec mode is not used in production, so
it is unclear when the functionality was lost. This makes the exec mode work
again.

Avoids the use of IO::Async in one instance and instead uses a tried and true
combination of system(), Try::Tiny, and Capture::Tiny.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>