Scan new format of md5sums file using unquoted file names, each terminanted by a null character.
This makes the collections less error prone for unusual file names. Was
inspired by #929729 but does not yet solve it.

The present approach will soon be replaced by a comprehensive solution. (The
'tar' utility offers --output-style literal but no null delimiter, so this
technique does not work for the index files.)

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>