Add test for repack indicator in Debian version.
Packages that were repacked usually include a Debian-specific string in the
upstream version. That can be something like 'debian' or 'dfsg' and triggers a
requirement to mangle the version in d/watch.

False positives can occur when these strings occur in the Debian revision. An
example is package 'elogin' in the Devuan distribution. This test memorializes
that particular case to avoid similar false positives in the future. For more
details, please see #931846.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>