Do not trigger repack requirement when the Debian revision includes the "repack" indicator. (Closes: #931846)
The Devuan distribution uses the string 'debian' in the Debian revision to
indicate whether a package was modified stock from Debian. That modification is
different from a source tarball that was repacked for Debian. It should not
trigger a repack requirement in debian/watch.

This commit fixes the regular expression per Mark Hindley's suggestion to avoid
such false positives in the future.

Signed-off-by: Chris Lamb <lamby@debian.org>