Add new module Lintian::Info::Changelog::Version; wraps version parsing.
This module makes a reliable mechanism to parse changelog versions
available to all checks (and the remainder of the Lintian code base).
Lintian previously used many different regular expressions. They were
consolidated here for consistency.

The only noteworthy item is that, for native packages, the upstream
portion always returns the empty string. That logic is different from
the wording in Debian's Policy but in the opinion of the author more
logical.

Raises an exception when parsing fails.

The full, unparsed version string is available via $version->literal.

Taken from checks/source-changelog.pm with some modifications.  For
example, character classes are specified with greater detail.

Gbp-Dch: ignore