In checks/watch-file.pm, use centralized version parser.
This is a better fix to Bug#931846 than the original patch. Now it is
much clearer to which part of the version string we are matching.

Now uses only the upstream part for pre-release detection. These
prerelease feature is probably not needed for native packages. They
are turned off for native packages.

Gbp-Dch: ignore