In Lintian::Collect::Changes, apply repack regex to locally upstream versions only.
It would be great to use the centralized version parser here, but the
native heuristics are only available in source packages.

Applies the original fix for #931846 to the regex locally.

Ideally, we would here rely on the 'repacked' method instead, if a
source package is mentioned in the changes file.

Gbp-Dch: ignore