Show logs for tests with parse errors when unattended and show all parse errors in TAP.
This should make it easier to diagnose build failures on Jenkins when lintian
crashes as sometimes happens for backports. Some issues, such as a missing
module or a function missing from the version of a module available in the
release target are not displayed clearly. Either error is evident only from
the individual test logs, which are inaccessible during unattended operation.

This commit prints the full log for each test with parse errors when
unattended. With this change, the overall log should be much clearer in case of
failure.

Also enables the option 'errors' in two TAP::Formatter classes to show all
parse errors under any circumstances. May result in the duplication of some
output.

Signed-off-by: Chris Lamb <lamby@debian.org>