Use non-native skeleton in test changelog-file-epoch-changed-upstream-version-increase for check changelog-file.
The tag epoch-changed-but-upstream-version-did-not-go-backwards applies only to
non-native packages but the test used the native skeleton. The discrepancy
showed up when version parsing was centralised and the tag disappeard. It now
uses the non-native skeleton to trigger the tag.

The version number was also native so we switch to a non-native version number.
It also adjusts the tags for the new version number.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>