Use non-native skeleton in test changelog-file-experimental for check changelog-file.
The tag non-consecutive-debian-revision is intended for non-native
packages, but the test for it uses a native skeleton. The
inconsistency showed up when version parsing was centralized, and the
tag disappeared. Instead uses the non-native skeleton to trigger the
non-native tag.

Also adjusts the expected tags for related changes.

Gbp-Dch: ignore