Always show log on failure to build a test package.
This makes it easier to debug unattended operations. It is possible, for
example, for a new package in unstable to affect the build negatively. The
error information is then available from the log in the build directory. Now we
print the log to stdout when there is a build failure.

Signed-off-by: Chris Lamb <lamby@debian.org>