Adjust systemd-general test for recent versions of debhelper. (Closes: #932339)
The upgrade of debhelper to version 12.2.2 exposed several issues in Lintian.

First of all, the systemd-general test had to be updated to use
dh_installsystemd which apparently did not exist when the test was written.
Second, it had to accommodate an update to dh_installinit which no longer
allows the naming pattern $daemon.init when '--name=$daemon' is used.
dh_installinit now requires the file to be called $package.$daemon.init. Some
other files were also renamed.

All of these issues are fixed here, except the symptom of hanging while
building test packages, which will be addressed separately in the solution to

This commit causes the unstable pipelines on jenkins and Gitlab to work again.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>