Again, fix pipeline for stretch-backports by using older process facility in IO::Async.
The open_process method to Loop was not available in IO::Async 0.71, which is
the version in stretch. This uses the conventional, and perhaps clearer, way of
explicitly creating an IO::Async::Process before adding it to the loop.

Addresses two more instances of open_process that were in the code base before
the changes to collections were accepted. The pipeline for stretch-backports
may have been broken for a little while.

Should fix the Jenkins pipeline for stretch-backports.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>