In checks/debian-source-dir.{desc,pm}, convert a tag to a non-recoverable program error.
The tag

  unknown-source-format

was found to be unreproducible using dpkg-source. It is considered
untestable in Lintian when using standard tools. The tag was converted
to a non-recoverable error that causes the program to exit.

The tag message is now an error message. If anyone ever sees it, let's
re-add the tag and use their package in the test suite.

Gbp-Dch: ignore