In checks/testsuite.{desc,pm}, convert three tags to non-recoverable program errors.
The three tags

  debian-tests-control-is-not-a-regular-file
  missing-runtime-tests-field
  syntax-error-in-debian-tests-control

were found to be unreproducible using dpkg-source. They are considered
untestable in Lintian when using standard tools. The tags were
converted to non-recoverable errors that cause the program to exit.

The tag messages are now error messages. If anyone ever sees them,
let's re-add the tags and use their packages in the test suite.

Gbp-Dch: ignore