Provide a separate module Test::Lintian::Build for the building of test packages.
This was previously part of the t/bin/build-test-packages but could in
the future be invoked in multiple ways. The Gitlab CI setup, for
example, may build in a different order than a developer at home.

Splits the code to build a test package into a separate module.

This commit is part of a series to allow adaptive and partial building
of test packages. It will ultimately solve Bug#935907.

Gbp-Dch: ignore