Add new tag 'empty-debian-tests-control'.
This tag replaces 'syntax-error-in-debian-tests-control' which was removed
recently because it was difficult to generate using standard tools, which
already caught those errors. The old tag was actually produced, perhaps
incorrectly, when d/tests/control contained only comments. The new tag now
indicates precisely that condition.

The severity of the tag was bumped to 'important' because the README for
autopkgtest says that one of the fields Tests: or Test-Command: is required.
Thanks to Vergo and pabs on #debian-mentors.

Gbp-Dch: ignore

Signed-off-by: Chris Lamb <lamby@debian.org>