Add module Lintian::Info::Changelog based on Parse::DebianChangelog.
In response to Bug#933134, the aging module Parse::DebianChangelog was
made a part of Lintian for customization and improvement. The Perl
team decided to retire it. This is the imported module.

Some improvements have already been made, and more are coming. The API
was simplified. Furthermore, the module will soon offer enhanced
changelog version parsing for convenient usage in Lintian checks.

Gbp-Dch: ignore