Drop Depends: and Build-Depends: on libparse-debianchangelog-perl. (Closes: #933134)
The module Parse::DebianChangelog is obsolete and was imported into
Lintian for customization and improvement. Alternatives such as
Dpkg::Changelog::Debian exist, but the effective interaction with
these datastructures is essential when writing checks.  Also, from a
risk perspective it probably makes sense for Lintian, a critic, to be
less dependent on the remaining Debian universe.

Removes the dependency on Parse::DebianChangelog.