Record check relationship in tag description.
A tag that is used in two checks could in the future cause the
issuance of a duplicate.  For that reason tags are, and always have
been, expected to be associated with a single check. This commit
records the relationship in a field in the tag description.

The field may hold only one value.

A future commit will remove the reverse relationship, which was a
cumbersome list prone to cause merge conflicts, from the check
description.

Please do not use the same tag in two check. It is not possible to
keep the heuristics in both checks, which are textually so distant,
balanced over time so that they never issue duplicates.  Instead,
please create a new tag.  The new tag can have a similar name.

Gbp-Dch: ignore