In Lintian::Tag::Info, offer method to load tag description from a path.
This converts the instantiator ->new to work without parameters and
instead offloads the burden of initializing all internal data
structures to another function called ->load, which accepts a file
path to the tag description.

This is part of an effort to collect more functionality related to
tags in the class Lintian::Tag::Info.

Gbp-Dch: ignore