In tags, quote spelling corrections that consist of multiple words.
When spelling corrections encompass several words, Lintian tags show
the correct and the incorrect spelling, but do not separate them in
any way. That can look confusing:

  I: mdadm: spelling-error-in-manpage usr/share/man/man8/mdadm.8.gz allows to allows one to

This commit quotes both the correct and the incorrect spelling. The
message then looks like:

  I: mdadm: spelling-error-in-manpage usr/share/man/man8/mdadm.8.gz "allows to" "allows one to"

The resulting messages are easier to read and understand.

The quotes do not appear when only single words are involved.