Mark the test 'changelog-file-become-native' is non-native in the changelog-file check.
Per the test's description, it is designed for when a maintainer
forgets to add Debian revision to an upstream version in the Debian
changelog. Such a package is always non-native, but the test used a
native skeleton.  Instead uses a non-native skeleton for the test.

The inconsistency showed up when version parsing was centralized.

Also adjusts the test's expected tags for related changes.

Gbp-Dch: ignore