Revert "Fix use of uninitialized value in checks/files.pm."
This reverts commit d55f56eca93e1e8ae288c593b3fbd0f9da31f092.

Commit d55f56ec broke the following semantics:

  $ perl -we 'my $p = "usr/sbin/foo"; print "/$1bin/$2\n" if $p =~ m,^(usr/)?sbin/(.+),;'
  /usr/bin/foo
  $ perl -we 'my $p = "sbin/foo"; print "/$1bin/$2\n" if $p =~ m,^(usr/)?sbin/(.+),;'
  Use of uninitialized value $1 in concatenation (.) or string at -e line 1.
  /bin/foo

So if $fname starts with usr/sbin/ (resp. sbin/) a corresponding name
with the /usr/bin/ (resp. /bin/) prefix is appended to @bin_binaries.
It also spews a warning for sbin/; I guess the desired outcome is to
silence that warning but preserve the overall behavior, right?  This
is no longer what happens: while names starting with usr/sbin/ are
dealt with like before, those starting with sbin/ are now ignored.

More information may be available here:

  https://salsa.debian.org/lintian/lintian/merge_requests/235#note_100475

Thanks to Guilhem Moulin for pointing it out!

Gbp-Dch: ignore