Attempt to fix irregular HTML display of package matlab-gdf in HTML output. (Re. #907425)
As lamby pointed out, the package matlab-gdf that gave rise to the bug
belongs to the section 'contrib/science'. The section is somewhat
unusual because it uses a slash.

No slash was present in the HTML page source (as one might expect for
an unescaped character) and the output looked a little different from
what was reported in the bug. (The version and the section were both
absent.)  Nonetheless, this patch seemed like a reasonable attempt to
resolve the bug. It is supposed to escape forward slashes to the HTML
sequence '&#x2f;' but it is untested.

It looks like the archive component for individual packages runs
through the routine html_quote, which is changed here, before being
passed to Text::Template:

https://salsa.debian.org/lintian/lintian/blob/master/commands/reporting-html-reports.pm#L1062

This commit may or may not close the bug. It needs to be tested.

Gbp-Dch: ignore