Fix instructions on how to run the test suite.
This does not bring the documentation in line with all recent changes,
but will hopefully avoid the worst vitriol from readers for its poor
state.

Fixes the instructions on how to run the test suite.

Also marks the tutorials for writing tests and checks as out of date.

Gbp-Dch: ignore