Generate HTML manual from reStructuredText; use latter as plain text. (Closes: #865847)
Lintian's manual was previously converted to reStructuredText format.
This commit uses rst2html to generate the HTML version.

RST does not have a facility to split documents into multiple output
files (although a 'toctree' directive could link multiple documents.)
For now, the HTML manual is one large, contiguous file.

Installs the single HTML file and the reStructuredText file in the
user package. Sets a link to the RST version with the ending *.txt in
hope of avoiding breakage.

Updates doc-base entries to point to the single HTML file and to the
RST file in lieu of plain text documentation.